+++ src/lib/jwt.ts
@@ -1,56 +1,56 @@
-// import { BadRequestException, Injectable, Module } from '@nestjs/common';
-// import { ConfigModule, ConfigService } from '@nestjs/config';
-// import { JwtModule, JwtModuleOptions, JwtOptionsFactory } from '@nestjs/jwt';
-// import { AuthGuard, PassportModule, PassportStrategy } from '@nestjs/passport';
-// import { ExtractJwt, Strategy } from 'passport-jwt';
+import { BadRequestException, Injectable, Module } from '@nestjs/common';
+import { ConfigModule, ConfigService } from '@nestjs/config';
+import { JwtModule, JwtModuleOptions, JwtOptionsFactory } from '@nestjs/jwt';
+import { AuthGuard, PassportModule, PassportStrategy } from '@nestjs/passport';
+import { ExtractJwt, Strategy } from 'passport-jwt';
 
-// @Injectable()
-// export class JwtStrategy extends PassportStrategy(Strategy) {
-//     constructor(private readonly configService: ConfigService) {
-//         super({
-//             jwtFromRequest: ExtractJwt.fromHeader('authorization'), // 헤더에 담겨서 넘어올 시 이름
-//             ignoreExpiration: false, // 만료된 토큰의 사용 여부
-//             secretOrKey: configService.get<JwtModuleOptions>('jwt')?.secret ?? configService.get('JWT_SECRET')
-//         });
-//     }
+@Injectable()
+export class JwtStrategy extends PassportStrategy(Strategy) {
+    constructor(private readonly configService: ConfigService) {
+        super({
+            jwtFromRequest: ExtractJwt.fromHeader('authorization'), // 헤더에 담겨서 넘어올 시 이름
+            ignoreExpiration: false, // 만료된 토큰의 사용 여부
+            secretOrKey: configService.get<JwtModuleOptions>('jwt')?.secret ?? configService.get('JWT_SECRET')
+        });
+    }
 
-//     // 토큰 검증
-//     async validate(payload: any) {
-//         if (!payload.userId || !payload.userLevel) throw new BadRequestException('JWT Token 검증에 실패하였습니다.');
+    // 토큰 검증
+    async validate(payload: any) {
+        if (!payload.userId || !payload.userLevel) throw new BadRequestException('JWT Token 검증에 실패하였습니다.');
 
-//         return {
-//             userId: payload.userId,
-//             userLevel: payload.userLevel
-//         };
-//     }
-// }
+        return {
+            userId: payload.userId,
+            userLevel: payload.userLevel
+        };
+    }
+}
 
-// @Injectable()
-// export class JwtConfigService implements JwtOptionsFactory {
-//     constructor(private readonly configService: ConfigService) {}
+@Injectable()
+export class JwtConfigService implements JwtOptionsFactory {
+    constructor(private readonly configService: ConfigService) {}
 
-//     createJwtOptions() {
-//         const jwtOption = this.configService.get<JwtModuleOptions>('jwt');
+    createJwtOptions() {
+        const jwtOption = this.configService.get<JwtModuleOptions>('jwt');
 
-//         if (!jwtOption) throw new Error('JWT config not found');
-//         if (!jwtOption.signOptions?.expiresIn) throw new Error('JWT signOptions.expiresIn not found');
+        if (!jwtOption) throw new Error('JWT config not found');
+        if (!jwtOption.signOptions?.expiresIn) throw new Error('JWT signOptions.expiresIn not found');
 
-//         jwtOption.secret ??= this.configService.get('JWT_SECRET');
+        jwtOption.secret ??= this.configService.get('JWT_SECRET');
 
-//         return jwtOption;
-//     }
-// }
+        return jwtOption;
+    }
+}
 
-// @Module({
-//     imports: [
-//         ConfigModule,
-//         PassportModule,
-//         JwtModule.registerAsync({ imports: [ConfigModule], inject: [ConfigService], useClass: JwtConfigService })
-//     ],
-//     providers: [JwtStrategy],
-//     exports: [JwtModule]
-// })
-// export class JwtAuthModule {}
+@Module({
+    imports: [
+        ConfigModule,
+        PassportModule,
+        JwtModule.registerAsync({ imports: [ConfigModule], inject: [ConfigService], useClass: JwtConfigService })
+    ],
+    providers: [JwtStrategy],
+    exports: [JwtModule]
+})
+export class JwtAuthModule {}
 
-// @Injectable()
-// export class JwtAuthGuard extends AuthGuard('jwt') {}
+@Injectable()
+export class JwtAuthGuard extends AuthGuard('jwt') {}
