+++ src/lib/jwt_guard.ts
@@ -1,5 +1,5 @@
-import { Injectable } from '@nestjs/common';
-import { AuthGuard } from '@nestjs/passport';
+// import { Injectable } from '@nestjs/common';
+// import { AuthGuard } from '@nestjs/passport';
 
-@Injectable()
-export class JwtAuthGuard extends AuthGuard('jwt') {}
+// @Injectable()
+// export class JwtAuthGuard extends AuthGuard('jwt') {}
